+++ database/migrations/2018_08_14_070738_create_containers_table.php
@@ -14,24 +14,24 @@ class CreateContainersTable extends Migration
     public function up()
     {
         Schema::create('containers', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('description')->nullable();
             $table->boolean('folders_enabled')->default(false);
 
-            $table->unsignedInteger('mobile_skin_id')->nullable();
+            $table->unsignedBigInteger('mobile_skin_id')->nullable();
             $table->foreign('mobile_skin_id')
                 ->references('id')
                 ->on('skins')
                 ->onDelete('set null');
 
-            $table->unsignedInteger('tablet_skin_id')->nullable();
+            $table->unsignedBigInteger('tablet_skin_id')->nullable();
             $table->foreign('tablet_skin_id')
                 ->references('id')
                 ->on('skins')
                 ->onDelete('set null');
 
-            $table->unsignedInteger('desktop_skin_id')->nullable();
+            $table->unsignedBigInteger('desktop_skin_id')->nullable();
             $table->foreign('desktop_skin_id')
                 ->references('id')
                 ->on('skins')
